ai-backend/src/database/requirements.txt
requests
aiohttp
python-dotenv
sqlalchemy
alembic
pydantic
pytest
pytest-asyncio
pytest-cov
black
flake8
mypy
types-requests